# Issue539 optimize bbm pure