--- a/dynamics/src/cgParametricMomentum.cpp
+++ b/dynamics/src/cgParametricMomentum.cpp
@@ -141,7 +141,7 @@ template <int CG> void CGParametricMomentum<CG>::CheckPeriodicity(CGVector<CG>& 
 {
     // the two segments bottom, right, top, left, are each processed in parallel
     for (size_t seg = 0; seg < smesh.periodic.size(); ++seg) {
-        // #pragma omp parallel for
+        //#pragma omp parallel for
         for (size_t i = 0; i < smesh.periodic[seg].size(); ++i) {
 
             const size_t ptype = smesh.periodic[seg][i][0];